util: add sql version of build_reference_cache()
Move this from sync_ref so later areas can use it.
