+++ src/sync_ref/tests.rs
@@ -11,6 +11,7 @@
 //! Tests for the sync_ref module.
 
 use super::*;
+use std::ops::DerefMut as _;
 use std::sync::Arc;
 
 /// Tests main().
@@ -116,8 +117,7 @@ reference_zipcounts = 'workdir/refs/irsz_count_20200717.tsv'
         buf,
         r#"sync-ref: downloading 'https://www.example.com/osm/data/irsz_count_20200717.tsv'...
 sync-ref: removing 'workdir/refs/irsz_count_20190717.tsv'...
-sync-ref: indexing hazszamok_20190511.tsv...
-sync-ref: indexing hazszamok_kieg_20190808.tsv...
+sync-ref: creating index...
 sync-ref: ok
 "#
     );
